adding routes to the website
