+++ package.json
@@ -10,8 +10,11 @@
     "normalize.css": "^8.0.1",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
+    "react-redux": "^7.1.1",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "2.1.8",
     "react-slick": "^0.25.2",
+    "redux": "^4.0.4",
     "slick-carousel": "^1.8.1"
   },
   "scripts": {
