--- src/App.css
@@ -1,26 +1,0 @@
-.App {
-}
-
-b{
-  font-weight: 700;
-}
-
-.golden{
-  color:var(--gold);
-}
-
-.on-construction {
-  display: flex;
-  padding: 0px 10%;
-  flex-wrap: wrap;
-  overflow: hidden;
-  padding: 100px;
-  /* background-color: var(--primary); */
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-.on-construction > p {
-  color: var(--gray);
-  text-align: center;
-}
