+++ src/components/Header/Header.css
@@ -14,7 +14,8 @@
 }
 
 .logo{
-  transition: width 300ms;
+  opacity: 1;
+  transition: opacity 300ms;
   cursor: pointer;
 }
