+++ src/components/Page1/Page1.jsx
@@ -1,8 +1,6 @@
 import React, { createRef, useState, useCallback, useRef } from "react";
 import "./Page1.css";
-
-import { useEventListener, useMobile } from "../../customHooks";
-
+import { useEventListener, useMobile } from "../../helpers/customHooks";
 import NumberOneIcon from "./NumberOneIcon";
 import ServicesIcon from "./ServicesIcon";
 import CollabIcon from "./CollabIcon";
