+++ src/index.css
@@ -81,3 +81,20 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
     monospace;
 }
+/* Change Autocomplete styles in Chrome*/
+input:-webkit-autofill,
+input:-webkit-autofill:hover, 
+input:-webkit-autofill:focus,
+textarea:-webkit-autofill,
+textarea:-webkit-autofill:hover,
+textarea:-webkit-autofill:focus,
+select:-webkit-autofill,
+select:-webkit-autofill:hover,
+select:-webkit-autofill:focus {
+  /* border: 1px solid green; */
+  /* -webkit-text-fill-color: var(--primary); */
+  -webkit-box-shadow: 0 0 0px 1000px #3AC6E6 inset;
+  color:var(--primary);
+  border-radius: 5px;
+  transition: background-color 500ms ease-in;
+}
