+++ src/screens/Societe.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+
+export default function Societe(){
+  return(
+    <div>
+      <p>sqdqsdsqd</p>
+    </div>
+  )
+}
